\name{nnNorm-internal}
\alias{denorm01}
\alias{denorm28}
\alias{norm01}
\alias{norm28}
\alias{tricube}
\title{Internal nnNorm functions and variables}
\description{
  Internal nnNorm functions and variables
}
\usage{
denorm01(Xn,mins,maxs)
denorm28(Xn,mins,maxs)
norm01(X) 
norm28(X) 
tricube(x) 

}
\details{
  These are not to be called directly by the user.
}
\keyword{internal}



